tests/requirements.txt
PyYAML>=6.0.0
tqdm>=4.66.2
requests>=2.31.0